+++ src/client/java/de/selectiverender/mixin/sodium/BlockOcclusionCacheMixin.java
@@ -1,6 +1,7 @@
 package de.selectiverender.mixin.sodium;
 
 import de.selectiverender.SelectiveRenderState;
+import de.selectiverender.SelectiveRenderSettings;
 import me.jellysquid.mods.sodium.client.render.chunk.compile.pipeline.BlockOcclusionCache;
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
@@ -21,7 +22,8 @@ abstract class BlockOcclusionCacheMixin {
                                                     CallbackInfoReturnable<Boolean> cir) {
         if (SelectiveRenderState.shouldRender(pos)
                 && !SelectiveRenderState.shouldRender(pos.offset(direction))) {
-            cir.setReturnValue(true);
+            cir.setReturnValue(SelectiveRenderSettings.boundaryMode()
+                    != SelectiveRenderSettings.BoundaryMode.CULLED);
         }
     }
 }